# Rewrite some regex conditions to use index-compatible conditions