--- a/influxql/ast.go
+++ b/influxql/ast.go
@@ -1241,8 +1241,8 @@ func (s *SelectStatement) RewriteFields(ic IteratorCreator) (*SelectStatement, e
 //
 // Conditions that can currently be simplified are:
 //
-//     - host ~= /^foo$/ becomes host = 'foo'
-//     - host ~! /^foo$/ becomes host != 'foo'
+//     - host =~ /^foo$/ becomes host = 'foo'
+//     - host !~ /^foo$/ becomes host != 'foo'
 //
 // Note: if the regex contains groups, character classes, repetition or
 // similar, it's likely it won't be rewritten. In order to support rewriting