###### Required for all non-trivial PRs
- [x] Rebased/mergable
- [x] Tests pass
- [x] CHANGELOG.md updated

Fixes #7486.

This PR adds support for replacing regexes with non-regex conditions
when possible. Currently the following regexes are supported:

 - `host =~ /^server1$/` will be converted into `host = 'server1'`
 - `host !~ /^server1$/` will be converted into `host != 'server1'`

*Note*: if the regex expression contains repetition, groupings or character classes, then rewriting may not take place. For example, the condition: `name =~ /^foo.$/` will not
be rewritten as it cannot be rewritten as an exact match.

Support for some of these cases may arrive in the future.

Regexes that can be converted into simpler expression will be able to
take advantage of the tsdb index, making them significantly faster.

#### Benchmarks

The following benchmarks run a query along the lines of:
 
```
SELECT count(value) FROM cpu WHERE host = 'server1'
``` 

for the non-regex ones, and: 

```
SELECT count(value) FROM cpu WHERE host =~ /^server1$/
```

 for the regex ones.

`master` 6fd74a6a

```
BenchmarkServer_Query_Count_Where_500-4          	    1000	   1392067 ns/op	  909527 B/op	     572 allocs/op
BenchmarkServer_Query_Count_Where_1K-4           	    1000	   1349177 ns/op	  909332 B/op	     571 allocs/op
BenchmarkServer_Query_Count_Where_10K-4          	    1000	   1564794 ns/op	  908493 B/op	     569 allocs/op
BenchmarkServer_Query_Count_Where_100K-4         	    1000	   2161792 ns/op	  906965 B/op	     566 allocs/op
BenchmarkServer_Query_Count_Where_Regex_500-4    	     100	  21228299 ns/op	 1717204 B/op	     642 allocs/op
BenchmarkServer_Query_Count_Where_Regex_1K-4     	     100	  22877996 ns/op	 1717317 B/op	     642 allocs/op
BenchmarkServer_Query_Count_Where_Regex_10K-4    	      50	  20535616 ns/op	 1717303 B/op	     642 allocs/op
BenchmarkServer_Query_Count_Where_Regex_100K-4   	     100	  18272231 ns/op	 1717068 B/op	     641 allocs/op
PASS
ok  	github.com/influxdata/influxdb/cmd/influxd/run	21.132s
```

This PR produces the following results:

```
BenchmarkServer_Query_Count_Where_500-4          	    1000	   1272334 ns/op	  909657 B/op	     572 allocs/op
BenchmarkServer_Query_Count_Where_1K-4           	    1000	   1316998 ns/op	  909420 B/op	     571 allocs/op
BenchmarkServer_Query_Count_Where_10K-4          	    1000	   1770097 ns/op	  908471 B/op	     569 allocs/op
BenchmarkServer_Query_Count_Where_100K-4         	    1000	   2368152 ns/op	  906952 B/op	     566 allocs/op
BenchmarkServer_Query_Count_Where_Regex_500-4    	    1000	   2053715 ns/op	  913136 B/op	     635 allocs/op
BenchmarkServer_Query_Count_Where_Regex_1K-4     	    1000	   2823591 ns/op	  913161 B/op	     635 allocs/op
BenchmarkServer_Query_Count_Where_Regex_10K-4    	     500	   2273258 ns/op	  913110 B/op	     635 allocs/op
BenchmarkServer_Query_Count_Where_Regex_100K-4   	    1000	   2327778 ns/op	  913072 B/op	     635 allocs/op
PASS
ok  	github.com/influxdata/influxdb/cmd/influxd/run	24.178s
```

So you can see that the regexes have been rewritten and are now effectively running the same benchmarks.

##### Ratings repo

A real world query on the [ratings](github.com/jwilder/ratings) dataset:

`1.0.2`
```
⇒  time influx -database ML1M -execute "SELECT mean(rating) FROM movielens WHERE userid =~ /^48$/"
name: movielens
---------------
time	mean
0	3.068561872909699

influx -database ML1M -execute   0.00s user 0.01s system 0% cpu 1.945 total
```

`PR`
```
⇒  time influx -database ML1M -execute "SELECT mean(rating) FROM movielens WHERE userid =~ /^48$/"
name: movielens
---------------
time	mean
0	3.068561872909699

influx -database ML1M -execute   0.00s user 0.01s system 0% cpu 1.399 total
```

@jwilder @jsternberg