@mark-rushakoff 2016-10-20
Any reason for not using [regexp/syntax](https://golang.org/pkg/regexp/syntax/) to inspect the regular expression, rather than inspecting the string? That would make it more straightforward to handle some of the other one-of-many-exact matches, e.g. `/^(foo|bar)$/`.