@e-dard 2016-10-21
@jwilder @jsternberg updated.

@mark-rushakoff I'll admit I'm pretty clueless about`regexp/syntax`. In the end I just went for a simple regex to identify strings like `^foo$` and then a simple rejection of any regex with a meta character. Can `regexp/syntax` do this for us in as simple a way?