@daviesalex 2016-10-21
This looks like a great idea to achieve a free lunch!

This may be out of the scope of this PR, but a really huge win here (that I briefly chatted to Paul about a few weeks ago) would be to add a more optimal "begins with" comparison, and filter those out. 99.9% of our regex's are host =~ /^server1/  style (may be expressed at host =~ /^server.*$/  or host =~ /^server.*/)

There are cases where the rex drops down to cases covered by this PR, but its not common (for us!)