@e-dard 2016-10-21
@daviesalex we have your use-case on our minds. I think the main problem preventing us solving the prefix issue with our current in-memory index, is that we can only do direct lookups for series keys. To do prefix lookups we would need to store series in-order.

The new `tsi` index approach over in https://github.com/influxdata/influxdb/tree/tsi could allow us to search through series keys in this way. Assuming that we add query support for prefix matches, we could then potentially rewrite `/^server1/`, `/server1./`, `/server1.+$` etc to use the index-compatible prefix search.