@jsternberg 2016-10-21
@e-dard `regexp/syntax actually seems to be pretty simple. I can work on it with you later today if you want. It would certainly be better and would allow us to support more types of optimizations such as character classes more easily.