@jsternberg 2016-10-21
Example of some code that I haven't finished yet:

```
func literalMatches(re *regexp.Regexp) ([]string, error) {
    s, err := syntax.Parse(re.String(), 0)
    if err != nil {
        return nil, err
    }

    // We are looking for EXACT matches which means we need
    // a concatenation of a ^, some text, then $.
    if s.Op != syntax.OpConcat && len(s.Sub) < 3 {
        return nil, nil
    }

    start := s.Sub[0]
    if !(start.Op == syntax.OpBeginLine || start.Op == syntax.OpBeginText) {
        return nil, nil
    }
    end := s.Sub[len(s.Sub)-1]
    if !(end.Op == syntax.OpEndLine || end.Op == syntax.OpEndText) {
        return nil, nil
    }

    // Look at the inside to try and put together a string.
    subs := s.Sub[1 : len(s.Sub)-1]
    if len(subs) == 0 {
        return []string{""}, nil
    }

    // now process the inside of the regular expression to try and find exact matches
    return nil, nil
}
```