@e-dard 2016-10-21
@jsternberg though it's important to remember that in the scope of this PR we can't support character classes, since they don't map to literal strings. In the future (beyond the scope of this PR) it would be cool if we could take `host =~ /^server-[ab]$/`  and convert that into `host = 'server-a' OR host = 'server-b'` 😃 .