@mark-rushakoff 2016-10-21
+1 for the code @jsternberg showed for this, and strong -1 from me on attempting to parse stringified regular expressions, with regular expressions. 

I think adding a check for `len(subs) == 1` and `subs[0].Op == syntax.OpLiteral` would more or less implement the issue as written. I think we'd want to run [(syntax.Regexp).Simplify](https://golang.org/pkg/regexp/syntax/#Regexp.Simplify) at the start of that code, too.

@stevenh linked to [this playground link](https://play.golang.org/p/EwEb9rXfX1) in a now-hidden comment, which seems like a good example of a fuller solution, perhaps for 1.2.