@stevenh 2016-10-21
I believe the original use case came from us identifying that grafana generates its queries that reference vars with regexp so you can't (from the gui) get:
```go
mytag = '$var'
```
instead you get:
```go
mytag = /^$var$/
```
However if $var is multiple then this results in:
```go
mytag =/^(value1|value2)$/
```
This seems like a simple case however syntax reduces this on parse to:
```go
mytag = /^value[1-2]$/
```

This is why I prototyped out a bit more in https://play.golang.org/p/EwEb9rXfX1 so we can actually deal with what grafana generates commonly.